fix recv memoryleak (#468)
- recv memory leak
- peek double free
- uninitialize malloc, change to calloc

+++ src/lava/magma/runtime/message_infrastructure/setenv.sh
@@ -1,7 +1,9 @@
 SCRIPTPATH=$(cd `dirname -- $BASH_SOURCE` && pwd)
 export MSG_LOG_PATH="${SCRIPTPATH}/log"
 if [ -d $MSG_LOG_PATH ]; then
-  rm $MSG_LOG_PATH/lava_message_infrastructure_pid_*.log
+  if [ -n "`find $MSG_LOG_PATH -maxdepth 1 -name 'lava_message_infrastructure_pid_*.log'`" ]; then
+    rm $MSG_LOG_PATH/lava_message_infrastructure_pid_*.log
+  fi
 else
   mkdir -p $MSG_LOG_PATH
 fi
